+++ client/src/containers/analysis-visualization/analysis-legend/contextual-legend-item/component.tsx
@@ -2,6 +2,7 @@ import { useCallback, useMemo } from 'react';
 
 import { useAppDispatch } from 'store/hooks';
 import { setLayer } from 'store/features/analysis/map';
+import { NUMBER_FORMAT } from 'utils/number-format';
 import LegendItem from 'components/legend/item';
 import LegendTypeBasic from 'components/legend/types/basic';
 import LegendTypeCategorical from 'components/legend/types/categorical';
@@ -35,7 +36,9 @@ const ContextualLegendItem = ({ layer }: ContextualLegendItemProps) => {
     const props = {
       items: layer.metadata.legend.items.map((item) => ({
         ...item,
-        label: item.label || `${item.value}`,
+        label:
+          item.label ||
+          `${!Number.isNaN(item.value) ? NUMBER_FORMAT(item.value as number) : item.value}`,
       })),
     };
     switch (layer.metadata.legend.type) {